refactor: profile imports in lib
