+++ packages/shared/src/lib/core/account/actions/createNewAccount.ts
@@ -1,13 +1,15 @@
+import { addEmptyAccountActivitiesToAllAccountActivities } from '@core/activity/stores'
 import { localize } from '@core/i18n'
-import { activeAccounts, addAccountPersistedDataToActiveProfile, addAccountToActiveAccounts } from '@core/profile'
 import { createAccount } from '@core/profile-manager'
+import {
+    activeAccounts,
+    addAccountPersistedDataToActiveProfile,
+    addAccountToActiveAccounts,
+} from '@core/profile/stores'
 import { get } from 'svelte/store'
-
 import { DEFAULT_SYNC_OPTIONS } from '../constants'
 import { IAccountState } from '../interfaces'
-
 import { buildAccountStateAndPersistedData } from './buildAccountStateAndPersistedData'
-import { addEmptyAccountActivitiesToAllAccountActivities } from '@core/activity/stores'
 
 export async function createNewAccount(name?: string, color?: string): Promise<IAccountState> {
     const account = await createAccount({
