+++ packages/shared/src/lib/core/account/actions/syncBalance.ts
@@ -1,7 +1,7 @@
+import { updateActiveAccount } from '@core/profile/stores'
+import { get } from 'svelte/store'
 import { getBalance } from '../api/getBalance'
 import { selectedAccount, updateSelectedAccount } from '../stores'
-import { updateActiveAccount } from '@core/profile'
-import { get } from 'svelte/store'
 
 export async function syncBalance(accountIndex: number): Promise<void> {
     const balances = await getBalance(accountIndex)
