+++ packages/shared/src/lib/core/activity/actions/linkTransactionsWithClaimingTransactions.ts
@@ -1,10 +1,10 @@
 import { IAccountState } from '@core/account'
 import { addClaimedActivity, claimedActivities } from '@core/activity/stores'
-import { activeProfileId } from '@core/profile'
+import { activeProfileId } from '@core/profile/stores'
 import { get } from 'svelte/store'
 import { ActivityDirection } from '../enums'
-import { isOutputAsync } from '../utils/outputs'
 import { IProcessedTransaction } from '../types'
+import { isOutputAsync } from '../utils/outputs'
 
 /**
  * It takes a list of transactions and links the transactions that are claiming async transactions
