+++ packages/shared/src/lib/core/activity/utils/generateSingleBasicActivity.ts
@@ -1,7 +1,8 @@
 import { isShimmerClaimingTransaction } from '@contexts/onboarding/stores'
 import { IAccountState } from '@core/account'
 import { IActivityGenerationParameters } from '@core/activity/types'
-import { activeProfileId, getCoinType } from '@core/profile'
+import { getCoinType } from '@core/profile/actions'
+import { activeProfileId } from '@core/profile/stores'
 import { IBasicOutput } from '@iota/types'
 import { get } from 'svelte/store'
 import { activityOutputContainsValue } from '..'
