+++ packages/shared/src/lib/core/activity/utils/generateSingleFoundryActivity.ts
@@ -1,10 +1,11 @@
 import { IAccountState } from '@core/account'
-import { getCoinType } from '@core/profile'
-import { ADDRESS_TYPE_ALIAS, UNLOCK_CONDITION_IMMUTABLE_ALIAS } from '@core/wallet/constants'
-import { ActivityType } from '../enums'
 import { IActivityGenerationParameters } from '@core/activity/types'
-import { FoundryActivity } from '../types'
+import { getCoinType } from '@core/profile/actions'
+import { ADDRESS_TYPE_ALIAS, UNLOCK_CONDITION_IMMUTABLE_ALIAS } from '@core/wallet/constants'
+import { convertHexAddressToBech32 } from '@core/wallet/utils'
 import type { IAliasAddress, IFoundryOutput, IImmutableAliasUnlockCondition } from '@iota/types'
+import { ActivityType } from '../enums'
+import { FoundryActivity } from '../types'
 import {
     getAmountFromOutput,
     getAsyncDataFromOutput,
@@ -13,7 +14,6 @@ import {
     getTagFromOutput,
 } from './helper'
 import { getNativeTokenFromOutput } from './outputs'
-import { convertHexAddressToBech32 } from '@core/wallet/utils'
 
 export function generateSingleFoundryActivity(
     account: IAccountState,
