+++ packages/shared/src/lib/core/activity/utils/outputs/getFormattedVotingPowerFromGovernanceActivity.ts
@@ -1,6 +1,6 @@
-import { getBaseToken } from '@core/profile'
 import { GovernanceAction } from '@core/activity/enums'
 import { GovernanceActivity } from '@core/activity/types'
+import { getBaseToken } from '@core/profile/actions'
 import { formatTokenAmountBestMatch } from '@core/wallet/utils'
 
 export function getFormattedVotingPowerFromGovernanceActivity(activity: GovernanceActivity): string {
