+++ packages/shared/src/lib/core/i18n/utils/formatCurrency.ts
@@ -1,7 +1,6 @@
-import { get } from 'svelte/store'
-
 import { appSettings } from '@core/app'
-import { activeProfile } from '@core/profile'
+import { activeProfile } from '@core/profile/stores'
+import { get } from 'svelte/store'
 
 export function formatCurrency(
     value: number | undefined,
