+++ packages/shared/src/lib/core/network/actions/addOfficialNodesToClientOptions.ts
@@ -1,5 +1,5 @@
 import { getOfficialNodes, updateClientOptions } from '@core/network'
-import { activeProfile } from '@core/profile'
+import { activeProfile } from '@core/profile/stores'
 import { get } from 'svelte/store'
 
 export function addOfficialNodesToClientOptions(): void {
