+++ packages/shared/src/lib/core/network/actions/toggleLocalPowInClientOptions.ts
@@ -1,5 +1,5 @@
 import { updateClientOptions } from '@core/network'
-import { activeProfile } from '@core/profile'
+import { activeProfile } from '@core/profile/stores'
 import { get } from 'svelte/store'
 
 export async function toggleLocalPowInClientOptions(): Promise<void> {
