+++ packages/shared/src/lib/core/network/actions/togglePrimaryNodeInClientOptions.ts
@@ -1,5 +1,5 @@
 import { INode, updateClientOptions } from '@core/network'
-import { activeProfile } from '@core/profile'
+import { activeProfile } from '@core/profile/stores'
 import { get } from 'svelte/store'
 
 export async function togglePrimaryNodeInClientOptions(node: INode): Promise<void> {
