+++ packages/shared/src/lib/core/nfts/utils/parseNftMetadata.ts
@@ -1,10 +1,10 @@
+import { getNetworkHrp } from '@core/profile/actions'
 import { Converter } from '@core/utils/convert'
 import { validateBech32Address } from '@core/utils/crypto'
 import { isValidUri } from '@core/utils/validation'
 import { Irc27Version, TokenStandard } from '@core/wallet/enums'
 import { IIrc27Attribute, IIrc27Metadata, ISoonaverseAttribute, ISoonaverseAttributes } from '../interfaces'
 import { MimeType } from '../types'
-import { getNetworkHrp } from '@core/profile'
 
 export function parseNftMetadata(metadata: string): IIrc27Metadata | undefined {
     try {
