+++ packages/shared/src/lib/core/profile/actions/active-accounts/updateActiveAccountPersistedData.ts
@@ -1,6 +1,6 @@
-import { selectedAccountIndex, updateSelectedAccount } from '@core/account/stores'
 import { IPersistedAccountData } from '@core/account/interfaces'
-import { updateAccountPersistedDataOnActiveProfile } from '@core/profile'
+import { selectedAccountIndex, updateSelectedAccount } from '@core/account/stores'
+import { updateAccountPersistedDataOnActiveProfile } from '@core/profile/stores'
 import { get } from 'svelte/store'
 
 export function updateActiveAccountPersistedData(
