+++ packages/shared/src/lib/core/profile/actions/active-profile/loadPersistedProfileIntoActiveProfile.ts
@@ -1,4 +1,5 @@
-import { profiles, setActiveProfile } from '@core/profile'
+import { setActiveProfile } from '@core/profile/actions'
+import { profiles } from '@core/profile/stores'
 import { get } from 'svelte/store'
 
 export function loadPersistedProfileIntoActiveProfile(profileId: string): void {
