+++ packages/shared/src/lib/core/profile/actions/active-profile/reflectLockedStronghold.ts
@@ -1,4 +1,4 @@
-import { activeProfile, clearTimeStrongholdLastUnlocked } from '@core/profile'
+import { activeProfile, clearTimeStrongholdLastUnlocked } from '@core/profile/stores'
 import { get } from 'svelte/store'
 
 export function reflectLockedStronghold(): void {
