+++ packages/shared/src/lib/core/profile/actions/active-profile/resetActiveProfile.ts
@@ -1,5 +1,6 @@
-import { activeProfile, activeProfileId, INITIAL_ACTIVE_PROFILE, IProfile } from '@core/profile'
 import { Platform } from '@core/app'
+import { INITIAL_ACTIVE_PROFILE, IProfile } from '@core/profile'
+import { activeProfile, activeProfileId } from '@core/profile/stores'
 import { get } from 'svelte/store'
 
 export function resetActiveProfile(): void {
