+++ packages/shared/src/lib/core/profile/actions/active-profile/setActiveProfile.ts
@@ -1,6 +1,6 @@
-import { activeProfile, INITIAL_ACTIVE_PROFILE, IPersistedProfile, IProfile } from '@core/profile'
-import { activeProfileId } from '@core/profile/stores'
 import { Platform } from '@core/app'
+import { INITIAL_ACTIVE_PROFILE, IPersistedProfile, IProfile } from '@core/profile'
+import { activeProfile, activeProfileId } from '@core/profile/stores'
 
 export function setActiveProfile(persistedProfile: IPersistedProfile): void {
     activeProfile?.set(<IProfile>{ ...INITIAL_ACTIVE_PROFILE, ...persistedProfile })
