+++ packages/shared/src/lib/core/profile/actions/active-profile/unlockStronghold.ts
@@ -1,7 +1,7 @@
+import { IProfile } from '@core/profile'
 import { setStrongholdPassword } from '@core/profile-manager'
-import { activeProfile, IProfile } from '@core/profile'
+import { activeProfile, setTimeStrongholdLastUnlocked } from '@core/profile/stores'
 import { get } from 'svelte/store'
-import { setTimeStrongholdLastUnlocked } from '@core/profile/stores'
 
 export async function unlockStronghold(password: string, profile: IProfile = get(activeProfile)): Promise<void> {
     const { isStrongholdLocked } = profile
