+++ packages/shared/src/lib/core/router/subrouters/login-router.ts
@@ -1,16 +1,16 @@
 import { isLatestStrongholdVersion } from '@core/app'
-import { activeProfile, ProfileType } from '@core/profile'
-import { get, writable } from 'svelte/store'
-
-import { Subrouter } from '../classes'
-import { LoginRoute } from '../enums'
-import { IRouterEvent } from '../interfaces'
-import { appRouter } from '../routers'
+import { ProfileType } from '@core/profile'
+import { activeProfile } from '@core/profile/stores'
 import features from '@features/features'
+import { get, writable } from 'svelte/store'
 import {
     UpdateStrongholdRouter,
     updateStrongholdRouter,
 } from '../../../../../../desktop/views/update-stronghold/update-stronghold-router'
+import { Subrouter } from '../classes'
+import { LoginRoute } from '../enums'
+import { IRouterEvent } from '../interfaces'
+import { appRouter } from '../routers'
 
 export const loginRoute = writable<LoginRoute>(null)
 export const loginRouter = writable<LoginRouter>(null)
