+++ packages/shared/src/lib/core/wallet/actions/send/sendOutputFromStardust.ts
@@ -1,10 +1,11 @@
 import { IAccountState } from '@core/account/interfaces'
 import { ledgerPreparedOutput } from '@core/ledger'
-import { checkActiveProfileAuth, getIsActiveLedgerProfile } from '@core/profile'
+import { checkActiveProfileAuth } from '@core/profile/actions'
+import { getIsActiveLedgerProfile } from '@core/profile/stores'
+import { resetSendFlowParameters } from '@core/wallet/stores'
 import { Output } from '@core/wallet/types'
 import { validateSendConfirmation } from '@core/wallet/utils'
 import { signAndSendStardustTransaction } from './signAndSendStardustTransaction'
-import { resetSendFlowParameters } from '@core/wallet/stores'
 
 export async function sendOutputFromStardust(
     output: Output,
